add new design

+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSDraggingEnumerator.cs
@@ -1,3 +1,3 @@
 namespace AppKit;
 
-public delegate void NSDraggingEnumerator(NSDraggingItem draggingItem, int idx, ref bool stop);
+public delegate void NSDraggingEnumerator(NSDraggingItem draggingItem, nint idx, ref bool stop);
